fix: PR #32 리뷰 반영 — @Modifying 캐시 정리, COUNT 타입 안전성, 참조수 답변 단위 집계
- incrementViewCount에 clearAutomatically/flushAutomatically 추가하여 L1 캐시 정합성 보장
- COUNT(DISTINCT cms.message_id)로 변경하여 답변 단위 참조수 집계 (청크 중복 제거)
- Projection/DTO의 viewCount/referenceCount/popularityScore를 int→long으로 변경 (BIGINT 안전)

+++ src/main/java/com/kdd/document/dto/PopularDocumentResponse.java
@@ -8,9 +8,9 @@ public record PopularDocumentResponse(
         Long documentId,
         String title,
         String category,
-        int viewCount,
-        int referenceCount,
-        int popularityScore,
+        long viewCount,
+        long referenceCount,
+        long popularityScore,
         LocalDateTime updatedAt
 ) {
     public static PopularDocumentResponse from(PopularDocumentProjection p) {
